add line numbering to the highlight plugin
This patch exposes two of the highlight libraries options: the
ability to enable line numbering and the variable used to control
the width of the line number column.

This involved changing the way the format plugin parses parameters.
It is now capable of accepting key/value pairs as well as the previous
simple ordered arguments it required.

This also includes an update to the documentation of the highlight
plugin to advertise this feature.